Initial split and updates to PatchMaker init.
