+++ ofrak_patch_maker/ofrak_patch_maker/patch_maker.py
@@ -4,11 +4,10 @@
 **Usage:**
 
 ```python
+tc = SOME_GNU_VERSION_ARM_Toolchain(...)
 known_symbols = {"memcpy": 0xdeadbeef}
 patch_maker = PatchMaker(
-    program_attributes=attr,
-    toolchain_config=tc_config,
-    toolchain_version=tc_version,
+    toolchain_config=tc
     platform_includes="../usr/include",
     base_symbols=known_symbols
 )
@@ -64,9 +63,7 @@ from ofrak_type.memory_permissions import MemoryPermissions
 class PatchMaker:
     def __init__(
         self,
-        program_attributes: ProgramAttributes,
-        toolchain_config: ToolchainConfig,
-        toolchain_version: ToolchainVersion,
+        toolchain: Toolchain,
         platform_includes: Optional[Iterable[str]] = None,
         base_symbols: Mapping[str, int] = None,
         build_dir: str = ".",
@@ -95,9 +92,7 @@ class PatchMaker:
         We should not raise exceptions in protected APIs. Protected programming interfaces should
         not be used external to this class. Use outside of the class at your own risk.
 
-        :param program_attributes: information about ISA/hardware
-        :param toolchain_config: information that will be translated to various flags
-        :param toolchain_version: used to derive the toolchain we'll be using
+        :param toolchain: a Toolchain instance with compile, link, assemble, etc. methods
         :param platform_includes: Additional include directories
         :param base_symbols: maps symbol name to effective address for patches
         :param build_dir: output directory for build artifacts
@@ -105,9 +100,7 @@ class PatchMaker:
         """
         self._platform_includes = platform_includes
         self.build_dir = build_dir
-        self._toolchain = self._get_toolchain(
-            program_attributes, toolchain_config, toolchain_version, logger=logger
-        )
+        self._toolchain = toolchain
 
         # String to file path of symbols.inc. This will be a build artifact.
         self._base_symbols: Dict[str, int] = {}
@@ -122,26 +115,6 @@ class PatchMaker:
             self._base_symbols.update(base_symbols)
 
         self.logger = logger
-
-    @staticmethod
-    def _get_toolchain(
-        program_attributes: ProgramAttributes,
-        toolchain_config: ToolchainConfig,
-        toolchain_version: ToolchainVersion,
-        logger: Union[logging.Logger, ModuleType] = logging,
-    ) -> Toolchain:
-        """
-        :param program_attributes: information about ISA/hardware
-        :param toolchain_config: information that will be translated to various flags
-        :param toolchain_version: used to derive the toolchain we'll be using
-        :param logger:
-
-        :return: A Toolchain matching the given arguments
-        """
-        toolchain_cls = toolchain_version.value
-        return toolchain_cls(
-            processor=program_attributes, toolchain_config=toolchain_config, logger=logger
-        )
 
     def _extract_symbols(self, path: str) -> Dict[str, int]:
         """
